test_case/requirement.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
dicttoxml==1.7.4
exceptiongroup==1.0.1
func-timeout==4.3.5
idna==3.4
iniconfig==1.1.1
lxml==4.9.1
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
xmljson==0.2.1
xmltodict==0.13.0
